Trait regular::accept::Accept

pub trait Accept {
    type Symbol;

    fn accept<I: IntoIterator<Item = Self::Symbol>>(self, string: I) -> bool;
}

Trait for objects that can decide whether or not a string belongs to some language that is represented by the object.

Required Associated Types

type Symbol

The type of symbols in the strings.

Required Methods

fn accept<I: IntoIterator<Item = Self::Symbol>>(self, string: I) -> bool

Return true if the given string is member of the language represented by this object.