Module regular::alphabet

Traits and implementations of generic sets of symbol, called alphabets in the context of languages and strings.

Structs

Boolean

Alphabet containing all booleans {True, False}.

FullRange

An object which can be turned into an alphabet which has all valid instances of Sym as members of the alphabet.

Unit

Alphabet containing single symbol of the unit type.

Traits

Alphabet

A set of symbols.

IntoAlphabet

Coversion into an alphabet.