Struct regular::prelude::DFABuilder

pub struct DFABuilder<A: Alphabet, S: DFAStorage<A> = DefaultDFAStorage<A, <A as Alphabet>::Symbol>> { /* fields omitted */ }

Builder for a DFA.

Methods

impl<A> DFABuilder<A> where
A: Alphabet + Debug + Clone,
A::Symbol: Eq + Hash + Debug + Clone,

pub fn new<I: IntoAlphabet<IntoAlpha = A, Symbol = A::Symbol>>(
alphabet: I
) -> Self

Create a new DFABuilder with the given alphabet.

impl<A, S> DFABuilder<A, S> where
S: DFAStorage<A>,
S::State: Ord,
A: Alphabet,

pub fn new_with_storage(storage: S) -> Self

Create a new DFABuilder with a customer storage backend.

pub fn new_state(&mut self) -> S::State

Record and return a new state.

pub fn transition(
&mut self,
from: S::State,
sym: A::Symbol,
to: S::State
) -> Result<(), Error>

Record and validate a new transition.

Error

This function will error if the symbol was not a member of the stated alphabet. This function will error if either the from or to state is not a valid state (valid states are only returned from calls of the add_state function).

pub fn transitions(
&mut self,
transitions: impl IntoIterator<Item = (S::State, A::Symbol, S::State)>
) -> Result<(), Error>

Record and validate multiple transitions.

Error

See the Error documentation of transition for ways that this function can fail.

pub fn accept_states(
&mut self,
accept_states: impl IntoIterator<Item = S::State>
) -> &mut Self

Add to the set of accept states.

pub fn start_state(&mut self, start: S::State) -> &mut Self

Set the starting state.

pub fn dead_state(&mut self, dead: Option<S::State>) -> &mut Self

Set the dead/error state.

pub fn build(self) -> Result<DFA<A, S>, Error>

Build the