Trait regular::prelude::DFAStorage

pub trait DFAStorage<A: Alphabet> {
    type State: Clone + Eq;

    // Required methods
    fn from_alphabet(alphabet: A) -> Self;
    fn alphabet(&self) -> &A;
    fn all_states(&self) -> Vec<Self::State>;
    fn all_transitions(&self) -> Vec<(Self::State, A::Symbol, Self::State)>;
    fn contains_state(&self, state: &Self::State) -> bool;
    fn transition(
        &self,
        current: Self::State,
        sym: A::Symbol
    ) -> Option<Self::State>;
    fn add_state(&mut self) -> Self::State;
    fn add_transition(
        &mut self,
        from: Self::State,
        sym: A::Symbol,
        to: Self::State
    );

    // Provided method
    unsafe fn transition_unchecked(
        &self,
        current: Self::State,
        sym: A::Symbol
    ) -> Self::State { ... }
}

Backend for the DFA struct.

Separating the two types allows for an optimized representation of the transition and state storage, which can be the largest part in most DFAs.

See the DefaultDFAStorage for a sensible default backend.

Required Associated Types

type State: Clone + Eq

Type representing a state of the DFA.

Required Methods

fn from_alphabet(alphabet: A) -> Self

Construct a new instance of this storage from the provided alphabet.

fn alphabet(&self) -> &A

Return a reference to the alphabet used by this DFA.

fn all_states(&self) -> Vec<Self::State>

Return a list of all the valid states of this DFA.

fn all_transitions(&self) -> Vec<(Self::State, A::Symbol, Self::State)>

Return a list of all transitions of this DFA.

fn contains_state(&self, state: &Self::State) -> bool

Return true if the given state is valid in this DFA.

fn transition(&self, current: Self::State, sym: A::Symbol) -> Option<Self::State>

Return Some(end) if there exists a transition from current to end via the given symbol.

fn add_state(&mut self) -> Self::State

Return a new unique state.

fn add_transition(&mut self, from: Self::State, sym: A::Symbol, to: Self::State)

Record the given transition.

Provided Methods

unsafe fn transition_unchecked( &self, current: Self::State, sym: A::Symbol ) -> Self::State

Return the resulting state from transitioning from the current state given the sym symbol.Alphabet

Safety

This version of the transition function is exposed so that implementations can bypass underlying safety checks when specific variants are upheld.

The current state must have been the result of a previous call to add_state.

The sym symbol must be a valid symbol contained in the alphabet associated with this DFA.