Struct regular::DFA

pub struct DFA<A: Alphabet, S: DFAStorage<A> = DefaultDFAStorage<A, <A as Alphabet>::Symbol>> { /* private fields */ }

A deterministic finite automaton.

Implementations

impl<A, S> DFA<A, S> where
S: DFAStorage<A>,
S::State: Ord,
A: Alphabet + Clone,

pub fn intersection<S2, S3>(
&self,
other: &DFA<A, S2>
) -> Result<DFA<A, S3>, Error> where
S::State: Hash,
S2: DFAStorage<A>,
S2::State: Ord + Hash,
S3: DFAStorage<A>,
S3::State: Ord,
A: PartialEq + Clone,
A::Symbol: Clone,

Construct a new DFA that accepts the regular language that is the intersection of the regular languages represented by this DFA and another DFA.

pub fn union<S2, S3>(&self, other: &DFA<A, S2>) -> Result<DFA<A, S3>, Error> where
S::State: Hash,
S2: DFAStorage<A>,
S2::State: Ord + Hash,
S3: DFAStorage<A>,
S3::State: Ord,
A: PartialEq + Clone,
A::Symbol: Clone,

Construct a new DFA that accepts the regular language that is the union of the regular languages represented by this DFA and another DFA.

pub fn difference<S2, S3>(
&self,
other: &DFA<A, S2>
) -> Result<DFA<A, S3>, Error> where
S::State: Hash,
S2: DFAStorage<A>,
S2::State: Ord + Hash,
S3: DFAStorage<A>,
S3::State: Ord,
A: PartialEq + Clone,
A::Symbol: Clone,

Construct a new DFA that accepts the regular language that is the difference between regular languages represented by this DFA and another DFA.

pub fn complement(&self) -> Self where
Self: Clone,

Construct a new DFA that accepts the regular language that is the complement of the regular language represented by this DFA.

impl<A, S> DFA<A, S> where
S: DFAStorage<A>,
S::State: Ord,
A: Alphabet + Clone,

pub fn accept_states(&self) -> &[S::State]

The states of DFA that will cause it to accept a string.

pub fn dead_state(&self) -> Option<&S::State>

An optional state that signals early termination of the DFA, used to represent an error condition.

pub fn start_state(&self) -> &S::State

The starting state of the DFA.

pub fn into_builder(self) -> DFABuilder<A, S>

Convert this DFA back into the DFABuilder form.

impl<A, S> DFA<A, S> where
S: DFAStorage<A>,
S::State: Ord,
A: Alphabet,

pub fn accept<I: IntoIterator<Item = A::Symbol>>(&self, string: I) -> bool

Accept or reject a string based on the content of this DFA.

This will immediately reject any string that contains a symbol that is not in the given alphabet.

pub unsafe fn accept_unchecked<I: IntoIterator<Item = A::Symbol>>(
&self,
string: I
) -> bool

Accept or reject a string based on the content of this DFA, without performing checks abouts the validity of the string.

Safety

Must assert that all symbol values are contained in the alphabet prior to calling this function.