Trait regular::util::Step

pub unsafe trait Step: Clone + PartialOrd + Sized {
    fn steps_between(start: &Self, end: &Self) -> Option<usize>;

    fn forward(&self, count: usize) -> Option<Self>;

    fn backward(&self, count: usize) -> Option<Self>;

    fn successor(&self) -> Self { ... }

    fn successor_saturating(&self) -> Self { ... }

    unsafe fn successor_unchecked(&self) -> Self { ... }

    fn predecessor(&self) -> Self { ... }

    fn predecessor_saturating(&self) -> Self { ... }

    unsafe fn predecessor_unchecked(&self) -> Self { ... }
}

TODO(replace copied code): This is a copy of the Step trait from https://github.com/rust-lang/rust/pull/62886 When the PR is merged or when the trait hits stable, replace this.

Objects that have a notion of successor and predecessor.

The successor operation moves towards values that compare greater. The predecessor operation moves towards values that compare lesser.

Safety

This trait is unsafe because its implementation must be correct for the safety of unsafe trait TrustedLen implementations, and the results of using this trait can be otherwise trusted by unsafe code.

Required methods

fn steps_between(start: &Self, end: &Self) -> Option<usize>

Returns the number of successor steps needed to get from start to end.

Returns None if that number would overflow usize (or is infinite, or if end would never be reached).

Invariants

For any a, b, and n:

steps_between(&a, &b) == Some(n) if and only if a.forward(n) == Some(b)

steps_between(&a, &b) == Some(n) if and only if b.backward(n) == Some(a)

steps_between(&a, &b) == Some(n) only if a <= b

Corrolary: steps_between(&a, &b) == Some(0) if and only if a == b

Note that a <= b does not imply steps_between(&a, &b) != None; this is the case when it would take more than usize::MAX steps to get to b

steps_between(&a, &b) == None if a > b

fn forward(&self, count: usize) -> Option<Self>

Returns the value that would be obtained by taking the successor of self count times.

Returns None if this would overflow the range of values supported by Self.

Invariants

For any a, n, and m where n + m does not overflow:

a.forward(n).and_then(|x| x.forward(m)) == a.forward(n + m)

a.forward(n) equals Step::successor applied to a n times

Corollary: a.forward(0) == Some(a)

a.forward(n).unwrap() >= a

fn backward(&self, count: usize) -> Option<Self>

Returns the value that would be obtained by taking the predecessor of self count times.

Returns None if this would underflow the range of values supported by Self.

Invariants

For any a, n, and m where n + m does not overflow:

a.backward(n).and_then(|x| x.backward(m)) == a.backward(n + m)

a.backward(n) equals Step::predecessor applied to a n times

Corollary: a.backward(0) == Some(a)

a.backward(n).unwrap() <= a

Loading content...
Provided methods

fn successor(&self) -> Self

Returns the successor of self.

If this would overflow the range of values supported by Self, this method is allowed to panic or wrap. Suggested behavior is to panic when debug assertions are enabled, and wrap otherwise.

Invariants

For any a where a.successor() does not overflow:

a == a.successor().predecessor()

a.successor() == a.forward(1).unwrap()

a.successor() >= a

fn successor_saturating(&self) -> Self

Returns the successor of self.

If this would overflow the range of values supported by Self, this method is defined to return the input value instead.

Invariants

For any a where a.successor() does not overflow:

a == a.successor().predecessor()

a.successor() == a.forward(1).unwrap()

a.successor() >= a

For any a where a.successor() does overflow:

a.successor() == a

unsafe fn successor_unchecked(&self) -> Self

Returns the successor of self without overflow.

Safety

It is undefined behavior if this operation exceeds the range of values supported by Self. If you cannot guarantee that this will not overflow, use forward or successor instead.

For any a, if there exists b such that b > a, it is safe to call a.successor_unchecked().

fn predecessor(&self) -> Self

Returns the predecessor of self.

If this would underflow the range of values supported by Self, this method is allowed to panic or wrap. Suggested behavior is to panic when debug assertions are enabled, and wrap otherwise.

Invariants

For any a where a.predecessor() does not underflow:

a == a.predecessor().successor()

a.predecessor() == a.backward(1).unwrap()

a.predecessor() <= a

fn predecessor_saturating(&self) -> Self

Returns the predecessor of self.

If this would underflow the range of values supported by Self, this method is defined to return the input value instead.

Invariants

For any a where a.predecessor() does not underflow:

a == a.predecessor().successor()

a.predecessor() == a.backward(1).unwrap()

a.predecessor() <= a

For any a where a.predecessor() does underflow:

a.predecessor() == a

unsafe fn predecessor_unchecked(&self) -> Self

Returns the predecessor of self without underflow.

Safety

It is undefined behavior if this operation exceeds the range of values supported by Self. If you cannot guarantee that this will not underflow, use backward or predecessor instead.

For any a, if there exists b such that b < a, it is safe to call a.successor_unchecked().

Loading content...