Trait regular::alphabet::Alphabet

pub trait Alphabet {
    type Symbol;
    type ValueIter: Iterator<Item = Self::Symbol>;
    fn values(&self) -> Self::ValueIter;

    fn contains(&self, sym: &Self::Symbol) -> bool;

    fn num_values(&self) -> Option<usize>;
}

A set of symbols.

Associated Types

type Symbol

The type of elements in this set.

type ValueIter: Iterator<Item = Self::Symbol>

An iterator over all values in this alphabet.

Loading content...
Required methods

fn values(&self) -> Self::ValueIter

Return an iterator over all value in this alphabet.

fn contains(&self, sym: &Self::Symbol) -> bool

Return true if the given symbol is a member of this alphabet.

fn num_values(&self) -> Option<usize>

Optionally return the number of elements in this alphabet. Return None if the size is unbounded or would overflow a usize value.

Loading content...